chore: Reorganize into library to make debugging easier and add new prepare method.
