+++ sample_config.py
@@ -3,6 +3,8 @@ import logging
 user_name = ""
 bot_password = ""
 user_name_only = ""
-user_agent = f"lexdannet, see https://github.com/dpriskorn/lexdannet User:{user_name_only}"
+user_agent = (
+    f"lexdannet, see https://github.com/dpriskorn/lexdannet User:{user_name_only}"
+)
 
 loglevel = logging.ERROR
