+++ src/models/form.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Form(BaseModel):
+    form: str
+    id: str
