+++ src/models/part_of_speech.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class PartOfSpeech(BaseModel):
+    pos: str
+    id: str
